+++ react/components/SKUSelector/components/SKUSelector.tsx
@@ -245,11 +245,16 @@ const variationNameToDisplayVariation =
       .filter(Boolean) as DisplayOption[]
 
     if (sortVariationsByLabel) {
-      options.sort((a, b) => {
+      const allNumbers = options.every(
+        (option: any) => !Number.isNaN(option.label)
+      )
+      options.sort((a: any, b: any) => {
+        if (allNumbers) {
+          return a.label - b.label
+        }
         return a.label < b.label ? -1 : a.label > b.label ? 1 : 0
       })
     }
-
     return { name, originalName, options }
   }