+++ react/components/Share/components/SocialButton.js
@@ -89,6 +89,7 @@ export default class SocialButton extends Component {
         )}
         media={imageUrl}
         onClick={() => this.props.sendShareEvent(socialEnum)}
+        openShareDialogOnClick
         {...additionalProps}
       >
         {icon}